(#1) Add POC for alpine docker image
This commit adds a proof of concept dockerfile for building
an Alpine-linux based image for the Puppet language server.
It leverages the puppet-in-docker image to simply add the
languageserver per instructions into the container. This
should be cleaned up in a future commit to pull only the
strictly necessary files instead of cloning the repo and
running the language server from within that clone.